# examples/axum-example/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/error": {
      "get": {
        "operationId": "error_endpoint",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/error/cookie-jar-logout": {
      "post": {
        "operationId": "cookie_jar_endpoint",
        "tags": [
          "error"
        ],
        "description": "Logout endpoint that uses CookieJar to clear cookies",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "null"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/error/error-with-status": {
      "get": {
        "operationId": "error_endpoint_with_status_code",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/error/error-with-status2": {
      "get": {
        "operationId": "error_endpoint_with_status_code2",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse2"
                }
              }
            }
          },
          "404": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse2"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse2"
                }
              }
            }
          }
        }
      }
    },
    "/error/error2": {
      "get": {
        "operationId": "error_endpoint2",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse2"
                }
              }
            }
          }
        }
      }
    },
    "/error/header-map": {
      "get": {
        "operationId": "header_map_endpoint",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {},
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse2"
                }
              }
            }
          }
        }
      }
    },
    "/error/header-map2": {
      "get": {
        "operationId": "header_map_endpoint2",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {},
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse2"
                }
              }
            }
          }
        }
      }
    },
    "/error/status-code/{id}": {
      "delete": {
        "operationId": "status_code_endpoint",
        "tags": [
          "error"
        ],
        "description": "Delete endpoint that returns just a StatusCode",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Error response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/memos": {
      "post": {
        "operationId": "create_memo",
        "description": "Create a new memo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMemoRequest"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "update_memo",
        "description": "Update a memo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMemoRequest"
                }
              }
            }
          }
        }
      }
    },
...
    "/memos/{id}": {
      "get": {
        "operationId": "get_memo",
        "description": "Get memo by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/memos/{id}/detail": {
      "get": {
        "operationId": "get_memo_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/memos/{id}/rel": {
      "get": {
        "operationId": "get_memo_rel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoResponseRel"
                }
              }
            }
          }
        }
      }
    },
    "/memos/{id}/summary": {
      "get": {
        "operationId": "get_memo_summary",
        "description": "Get a memo summary (same-file relation adapter with a custom schema name)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMemoRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint32"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "code"
        ]
      },
      "ErrorResponse2": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint32"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "code"
        ]
      },
...
      "MemoCommentInMemoDetail": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "memoId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "id",
          "memoId",
          "content"
        ]
      },
...
      "MemoDetailResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "1970-01-01T00:00:00+00:00"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "memoComments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoCommentInMemoDetail"
            }
          },
          "status": {
            "$ref": "#/components/schemas/MemoStatus"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "default": "1970-01-01T00:00:00+00:00"
          },
          "user": {
            "$ref": "#/components/schemas/UserInMemoDetail",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "content",
          "status",
          "createdAt",
          "updatedAt",
          "user",
          "memoComments"
        ]
      },
      "MemoResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "1970-01-01T00:00:00+00:00"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "status": {
            "$ref": "#/components/schemas/MemoStatus"
          },
          "title": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "content",
          "status",
          "createdAt"
        ]
      },
...
      "MemoResponseRel": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "1970-01-01T00:00:00+00:00"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "status": {
            "$ref": "#/components/schemas/MemoStatus"
          },
          "title": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "content",
          "status",
          "createdAt",
          "user"
        ]
      },
...
      "MemoStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ]
      },
      "MemoSummaryResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "1970-01-01T00:00:00+00:00"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "note": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MemoStatus"
          },
          "title": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MemoSummaryUser",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "content",
          "status",
          "createdAt",
          "user",
          "note"
        ]
      },
      "MemoSummaryUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "UpdateMemoRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content",
          "id"
        ]
      },
...
      "UserInMemoDetail": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
...
      "UserSchema": {
        "type": "object",
        "description": "User model",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Created at",
            "default": "1970-01-01T00:00:00+00:00"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "User ID",
            "default": 0
          },
          "name": {
            "type": "string",
            "description": "User name"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at",
            "default": "1970-01-01T00:00:00+00:00"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  },
...
}