# examples/axum-example/src/routes/error.rs
use serde::{Deserialize, Serialize};
use vespera::{
    Schema,
    axum::{Json, http::StatusCode, http::header::HeaderMap, response::IntoResponse},
    axum_extra::extract::cookie::CookieJar,
};

#[derive(Serialize, Deserialize, Schema)]
pub struct ErrorResponse {
    pub error: String,
    pub code: u32,
}

#[derive(Serialize, Deserialize, Schema)]
pub struct ErrorResponse2 {
    pub error: String,
    pub code: u32,
}

impl IntoResponse for ErrorResponse2 {
    fn into_response(self) -> vespera::axum::response::Response {
        (StatusCode::INTERNAL_SERVER_ERROR, Json(self)).into_response()
    }
}

#[vespera::route()]
pub async fn error_endpoint() -> Result<&'static str, Json<ErrorResponse>> {
    Err(Json(ErrorResponse {
        error: "Internal server error".to_string(),
        code: 500,
    }))
}

#[vespera::route(path = "/error-with-status")]
pub async fn error_endpoint_with_status_code()
-> Result<&'static str, (StatusCode, Json<ErrorResponse>)> {
    Err((
        StatusCode::INTERNAL_SERVER_ERROR,
        Json(ErrorResponse {
            error: "Internal server error".to_string(),
            code: 500,
        }),
    ))
}

#[vespera::route(path = "/error2")]
pub async fn error_endpoint2() -> Result<&'static str, ErrorResponse2> {
    Err(ErrorResponse2 {
        error: "Internal server error".to_string(),
        code: 500,
    })
}

#[vespera::route(path = "/error-with-status2", error_status = [500, 400, 404])]
pub async fn error_endpoint_with_status_code2() -> Result<&'static str, (StatusCode, ErrorResponse2)>
{
    Err((
        StatusCode::INTERNAL_SERVER_ERROR,
        ErrorResponse2 {
            error: "Internal server error".to_string(),
            code: 500,
        },
    ))
}

#[vespera::route(path = "/header-map")]
pub async fn header_map_endpoint() -> Result<(HeaderMap, &'static str), ErrorResponse2> {
    let headers = HeaderMap::new();
    println!("headers: {:?}", headers);
    Ok((headers, "ok"))
}

#[vespera::route(path = "/header-map2")]
pub async fn header_map_endpoint2() -> Result<(StatusCode, HeaderMap, &'static str), ErrorResponse2>
{
    let headers = HeaderMap::new();
    println!("headers: {:?}", headers);
    // Success branch returns 200 (the generated spec infers 200 for the Ok
    // arm); returning 500 here was a fixture quirk that contradicted the spec.
    Ok((StatusCode::OK, headers, "ok"))
}

/// Delete endpoint that returns just a StatusCode
#[vespera::route(delete, path = "/status-code/{id}", tags = ["error"])]
pub async fn status_code_endpoint(
    vespera::axum::extract::Path(id): vespera::axum::extract::Path<i64>,
) -> Result<StatusCode, (StatusCode, String)> {
    if id == 0 {
        return Err((StatusCode::NOT_FOUND, "Not found".to_string()));
    }
    Ok(StatusCode::NO_CONTENT)
}

/// Logout endpoint that uses CookieJar to clear cookies
#[vespera::route(post, path = "/cookie-jar-logout", tags = ["error"])]
pub async fn cookie_jar_endpoint(
    jar: CookieJar,
) -> Result<(CookieJar, Json<()>), (StatusCode, String)> {
    Ok((jar, Json(())))
}